quantized data into 8 bit by compressing. Good thing we don't lose information bad news we didn't gain any performance boosts. At this point this guarantees that the bottleneck is the number of triangles we emit which is not really solvable as it totally depends on the data and we don't have a big world or smth so that we could cull

+++ src/Scenes/OrganVisualizationScene.h
@@ -23,7 +23,7 @@ public:
 private:
 	void loadData(uint32_t organID);
 	std::pair<std::vector<float>, glm::uvec3> loadCTheadData() const;
-	std::pair<std::vector<float>, glm::uvec3> loadOrganAtlasData(const char* organPathBase) const;
+	std::pair<std::vector<uint8_t>, glm::uvec3> loadOrganAtlasData(const char* organPathBase) const;
 	void clearBuffers();
 
 	void createChunkVisualizationBuffer(const std::vector<VolumeChunk>& chunks);
